requirements.txt
openai
sounddevice
python-dotenv
scipy
faster-whisper